1. Changes related to porting to 64 bit 2. Generalized unsigned int to unitptr_t (including in generated code) or OBJECT_PTR 3. Removed current directory prefix from .so file name in unit_tests.lisp

+++ src/native_compiler.c
@@ -183,7 +183,7 @@ extern OBJECT_PTR CONS_HALT_NIL;
 extern OBJECT_PTR TRUE;
 #endif
 
-extern unsigned int POINTER_MASK;
+extern uintptr_t POINTER_MASK;
 
 extern OBJECT_PTR first(OBJECT_PTR);
 extern OBJECT_PTR second(OBJECT_PTR);
